reference_code/in-vehicle/requirements.txt
paho-mqtt==1.6.1
requests==2.27.1
